% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_tmax}
\alias{daily_tmax}
\title{Climatic dataset}
\format{numeric vector with maximum temperatures in celsius degrees.
\describe{
  \item{\code{daily_tmax}}{numeric. daily maximum temperatures}
}}
\usage{
daily_tmax
}
\description{
Dataset with 30 years of daily maximum  temperature.
}
\details{
Daily maximum  temperature
}
\keyword{datasets}
