% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_tmin}
\alias{daily_tmin}
\title{Climatic dataset}
\format{numeric vectors with minimumtemperatures in celsius degrees.
\describe{
  \item{\code{daily_tmin}}{numeric. daily minimum temperatures}
}}
\usage{
daily_tmin
}
\description{
Dataset with 30 years of daily minimum temperature.
}
\details{
Daily minimum temperature
}
\keyword{datasets}
