% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstFrost.R
\name{firstFrost}
\alias{firstFrost}
\title{Mean first frost day}
\usage{
firstFrost(
  mn,
  dates,
  iniday = "07-01",
  endday = "06-30",
  type = "doy",
  thres = 0
)
}
\arguments{
\item{mn}{vector of daily minimum temperature.}

\item{dates}{vector of dates corresponding to the mn vector, in Date format.}

\item{iniday}{first day of the growing season, in text format ("mm-dd").
Defaults to `07-01`.}

\item{endday}{last day of the growing season, in text format ("mm-dd").
Defaults to `06-30`.}

\item{type}{character vector specifying the type of output desired. It can
be `doy` for day of the year (julian day), or `date` for character date 
format (`mm-dd`). Defaults to `doy`.}

\item{thres}{temperature threshold to trigger frost (defaults to 0ºC).}
}
\value{
Depending on `type`, the output will be a numeric vector of julian
days (of `type` = 'doy') or a vector of characters representing dates
(`type` = 'date').
}
\description{
Calculates the first frost day (tmp < 0ºC) of the growing
season.
}
\details{
The function is able to span over consecutive years. If
`enddate` is an earlier day than `inidate`, then the considered period will
leap over two years.
}
\examples{

firstFrost(mn = daily_tmin,
            dates = seq.Date(as.Date('1981-01-01'),
                             as.Date('2010-12-31'), by ='day'),
            iniday = '07-01',
            endday = '06-30',
            type = 'doy')

}
