% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_plots.R
\name{overlap_plots}
\alias{overlap_plots}
\title{Overlap plots}
\usage{
overlap_plots(base_plt, over_plt)
}
\arguments{
\item{base_plt}{base plot.}

\item{over_plt}{Overlapping plot.}
}
\description{
This function overlay two plots build with ggplot2 and use both legends. This function was extracted from: https://stackoverflow.com/questions/40901169/ggplot2-overlay-two-plots-and-use-both-legends
}
\examples{

# Examples can be found at:
# https://stackoverflow.com/questions/40901169/ggplot2-overlay-two-plots-and-use-both-legends

}
\author{
"Constantine" and "Sandy Muspratt" users from StackOverflow
}
